Updating Patched
Another two bugs has been found and squished; these went all the way
down to the EventObject class.

1) When adding an EventListener to the EventObject's list you could pass
and add a null value, resulting an a crash any time you got the attached
EventListeners. This one was as simple as an existence check before
adding to the list of listeners.

2) This one was slightly trickier because of generic types. When getting
the EventListeners on an EventObject it casts the return value to an
array of your specified type. The issue was that the method to get the
ArrayList of listeners as an array it returned an Object[] which cannot
be cast to a generic array of T[]; the result of this is that i had to
use all sorts of black magic so that i could get an actual Class type to
cast to. But it has been done.