Pretty Colours, MASSIVELY Improved Rendering and Cleaned up Update Events
As a test for the tile system and the graphics i had the the tiles
display as a cascade of pretty colours. This created our favorite
flickering issue but i managed to finally get buffer strategies working
for me so now we are able to have multi threaded, back buffered,
hardware accelerate graphics with minimal effort.

There were a few bugs where i hadn't put in the code to add objects as
event listeners but this was mainly only an issue when it came to doing
a restart of the application from the commandline. Regardless i fixed up
the code and now it all works great, no memory leaks no issues.

Additionally to all this i updated the Logger so that we are able to
switch to a debug mode where everything that is written to the Logger is
printed on the console (also added more calls to write to the Logger) as
well as adding in a method for writing stack traces to the Logger in a
readable format.